Create break_up_large_string_for_google_trends
If you have ever worked with the google GCP translate V2 api to translate large amount of texts, then you are probably familiar with 429, or 501 errors telling you that you have sent a bad request, or the service is not available. As you probably know, that is because of the internal limits to how large a text you can send to the translate API (non-commercial) at once. I encountered this a lot when I was translating full Wikipedia pages. To bypass this, I wrote this program, which gets the length of the string you are trying to push, and then breaks it up by 5% steps and then feeds it piecewise into the translate api, then puts it all together again. Feel free to use it :D May it be a light to you in dark places when all other lights go out.<3